Create a custom hook used to encapsulate the update functionality (#2814)
- Apply to both participant and accelerator deliverable variable update
views
- Get rid of superfluous `DeliverableEditVariable` component
