+++ src/scenes/DeliverablesRouter/QuestionsDeliverableEditView.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useCallback, useEffect } from 'react';
 
 import { Box, Grid, useTheme } from '@mui/material';
 
@@ -7,46 +7,28 @@ import DeliverableVariableDetailsInput from 'src/components/DocumentProducer/Del
 import Card from 'src/components/common/Card';
 import WrappedPageForm from 'src/components/common/PageForm';
 import useNavigateTo from 'src/hooks/useNavigateTo';
+import { useProjectVariablesUpdate } from 'src/hooks/useProjectVariablesUpdate';
 import { useDeliverableData } from 'src/providers/Deliverable/DeliverableContext';
-import { selectUpdateVariableValues } from 'src/redux/features/documentProducer/values/valuesSelector';
-import {
-  requestListDeliverableVariablesValues,
-  requestUpdateVariableValues,
-} from 'src/redux/features/documentProducer/values/valuesThunks';
+import { requestListDeliverableVariablesValues } from 'src/redux/features/documentProducer/values/valuesThunks';
 import { selectDeliverableVariablesWithValues } from 'src/redux/features/documentProducer/variables/variablesSelector';
 import { requestListDeliverableVariables } from 'src/redux/features/documentProducer/variables/variablesThunks';
 import { useAppDispatch, useAppSelector } from 'src/redux/store';
-import strings from 'src/strings';
 import { VariableWithValues } from 'src/types/documentProducer/Variable';
-import {
-  NewDateValuePayload,
-  NewLinkValuePayload,
-  NewNumberValuePayload,
-  NewSelectValuePayload,
-  NewTextValuePayload,
-  Operation,
-  VariableValueDateValue,
-  VariableValueLinkValue,
-  VariableValueNumberValue,
-  VariableValueSelectValue,
-  VariableValueTextValue,
-  VariableValueValue,
-} from 'src/types/documentProducer/VariableValue';
-import useSnackbar from 'src/utils/useSnackbar';
+import { VariableValueValue } from 'src/types/documentProducer/VariableValue';
 
 const QuestionsDeliverableEditView = (): JSX.Element | null => {
   const dispatch = useAppDispatch();
   const theme = useTheme();
-  const snackbar = useSnackbar();
   const { goToDeliverable } = useNavigateTo();
   const { currentDeliverable: deliverable, deliverableId, projectId } = useDeliverableData();
 
-  const [pendingVariableValues, setPendingVariableValues] = useState<Map<number, VariableValueValue[]>>(new Map());
-  const [removedVariableValues, setRemovedVariableValues] = useState<Map<number, VariableValueValue>>(new Map());
-  const [updateVariableRequestId, setUpdateVariableRequestId] = useState<string>('');
-
   const variablesWithValues: VariableWithValues[] = useAppSelector((state) =>
     selectDeliverableVariablesWithValues(state, deliverableId, projectId)
+  );
+
+  const { pendingVariableValues, setRemovedValue, setValues, update, updateSuccess } = useProjectVariablesUpdate(
+    projectId,
+    variablesWithValues
   );
 
   useEffect(() => {
@@ -58,144 +40,23 @@ const QuestionsDeliverableEditView = (): JSX.Element | null => {
     void dispatch(requestListDeliverableVariablesValues({ deliverableId, projectId }));
   }, [deliverableId, projectId]);
 
-  const handleOnCancel = useCallback(() => {
+  const goToThisDeliverable = useCallback(() => {
     goToDeliverable(deliverableId, projectId);
   }, [deliverableId, projectId]);
 
-  const results = useAppSelector(selectUpdateVariableValues(updateVariableRequestId));
-
   useEffect(() => {
-    if (!results) {
-      return;
+    if (updateSuccess) {
+      goToThisDeliverable();
     }
-
-    if (results.status === 'success') {
-      snackbar.toastSuccess(strings.CHANGES_SAVED);
-      goToDeliverable(deliverableId, projectId);
-    } else if (results.status === 'error') {
-      snackbar.toastError(strings.GENERIC_ERROR);
-    }
-  }, [deliverableId, projectId, results]);
+  }, [goToThisDeliverable, updateSuccess]);
 
   const handleOnSave = () => {
     if (pendingVariableValues.size === 0) {
-      return;
+      // If the user clicks save but there are no changes, just navigate them back to the deliverable
+      goToThisDeliverable();
     }
 
-    const operations: Operation[] = [];
-
-    pendingVariableValues.forEach((pendingValues, variableId) => {
-      const variable = variablesWithValues.find((variableWithValues) => variableWithValues.id === variableId);
-      if (!variable) {
-        // TODO figure out error handling for this case
-        console.log('Variable not found');
-        return;
-      }
-
-      let newValue:
-        | NewDateValuePayload
-        | NewTextValuePayload
-        | NewNumberValuePayload
-        | NewSelectValuePayload
-        | NewLinkValuePayload
-        | undefined;
-
-      let newValues: NewTextValuePayload[] = [];
-      let valueIdToUpdate = -1;
-
-      if (variable.type === 'Text') {
-        const firstValue = pendingValues[0] as VariableValueTextValue;
-        valueIdToUpdate = firstValue.id;
-
-        newValues = pendingValues.reduce((acc: NewTextValuePayload[], nV: VariableValueValue) => {
-          if (nV.type === 'Text') {
-            acc.push({ type: 'Text', textValue: nV.textValue });
-          }
-          return acc;
-        }, []);
-      }
-
-      if (variable.type === 'Number') {
-        const firstValue = pendingValues[0] as VariableValueNumberValue;
-        valueIdToUpdate = firstValue.id;
-        newValue = { type: 'Number', numberValue: firstValue.numberValue, citation: firstValue.citation };
-      }
-
-      if (variable.type === 'Select') {
-        const firstValue = pendingValues[0] as VariableValueSelectValue;
-        valueIdToUpdate = firstValue.id;
-        newValue = { type: 'Select', optionIds: firstValue.optionValues, citation: firstValue.citation };
-      }
-
-      if (variable.type === 'Date') {
-        const firstValue = pendingValues[0] as VariableValueDateValue;
-        valueIdToUpdate = firstValue.id;
-        newValue = { type: 'Date', dateValue: firstValue.dateValue, citation: firstValue.citation };
-      }
-
-      if (variable.type === 'Link') {
-        const firstValue = pendingValues[0] as VariableValueLinkValue;
-        valueIdToUpdate = firstValue.id;
-        newValue = { type: 'Link', url: firstValue.url, citation: firstValue.citation, title: firstValue.title };
-      }
-
-      if (newValue) {
-        if (pendingValues[0].id !== -1) {
-          operations.push({
-            operation: 'Update',
-            valueId: valueIdToUpdate,
-            value: newValue,
-            existingValueId: valueIdToUpdate,
-          });
-        } else {
-          operations.push({ operation: 'Append', variableId: variable.id, value: newValue });
-        }
-      }
-
-      if (newValues) {
-        newValues.forEach((nV, index) => {
-          if (pendingValues[index].id !== -1) {
-            operations.push({
-              operation: 'Update',
-              valueId: pendingValues[index].id,
-              value: nV,
-              existingValueId: pendingValues[index].id,
-            });
-          } else {
-            operations.push({ operation: 'Append', variableId: variable.id, value: nV });
-          }
-        });
-
-        // delete list of values removed
-        const removedValue = removedVariableValues.get(variable.id);
-        if (removedValue) {
-          operations.push({
-            operation: 'Delete',
-            valueId: removedValue.id,
-            existingValueId: removedValue.id,
-          });
-        }
-      }
-    });
-
-    if (operations.length > 0) {
-      console.log({ operations });
-      const request = dispatch(
-        requestUpdateVariableValues({
-          operations,
-          projectId: projectId,
-        })
-      );
-      setUpdateVariableRequestId(request.requestId);
-    }
-  };
-
-  const setValues = (variableId: number, values: VariableValueValue[]) => {
-    setPendingVariableValues(new Map(pendingVariableValues).set(variableId, values));
-  };
-
-  const setRemovedValues = (variableId: number, value: VariableValueValue) => {
-    setRemovedVariableValues(new Map(removedVariableValues).set(variableId, value));
+    update();
   };
 
   if (!deliverable) {
@@ -205,7 +66,7 @@ const QuestionsDeliverableEditView = (): JSX.Element | null => {
   return (
     <WrappedPageForm
       cancelID={'cancelEditQuestionsDeliverable'}
-      onCancel={handleOnCancel}
+      onCancel={goToThisDeliverable}
       onSave={handleOnSave}
       saveID={'saveEditQuestionsDeliverable'}
     >
@@ -231,7 +92,7 @@ const QuestionsDeliverableEditView = (): JSX.Element | null => {
                       setValues={(newValues: VariableValueValue[]) => setValues(variableWithValues.id, newValues)}
                       variable={variableWithValues}
                       addRemovedValue={(removedValue: VariableValueValue) =>
-                        setRemovedValues(variableWithValues.id, removedValue)
+                        setRemovedValue(variableWithValues.id, removedValue)
                       }
                     />
                   </Grid>
